pyproject.toml
[project]
name = "ynab-agent"
version = "0.2.0"
description = "Agentic budget companion for YNAB via Claude Code skills"
requires-python = ">=3.12"
dependencies = [
    "ynab>=4.0.0",
    "python-dotenv>=1.1.0",
    "thefuzz>=0.22.0",
]

[project.scripts]
ynab-agent = "app.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=6.0.0",
    "genbadge[coverage]>=1.1.3",
    "ruff>=0.14.0",
    "ty>=0.0.8",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty.environment]
python-version = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
